pyproject.toml
[project]
name = "ollama-tuner"
version = "0.1.0"
description = "Fine-tune Ollama models on markdown files"
authors = [{name = "Stephen"}]
dependencies = [
    "ollama>=0.1.7",
    "aiofiles>=23.1.0", 
    "click>=8.1.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0"
]
requires-python = ">=3.8"

[project.scripts]
ollama-tuner = "src.main:cli"

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"